requirements.txt
Django>=4.2
djangorestframework
drf-spectacular
psycopg2-binary
celery[redis]
django-environ
gunicorn
